==> notebooks/1_query_database.ipynb <==
from pymongo import MongoClient
import pandas as pd

# Connecting to MongoDB in order to make the queries

client =  MongoClient("localhost:27017")

client.list_database_names()

db = client.get_database("ironhack")

# Connecting to the collection "companies", from which I will filter

companies = db.get_collection("companies")

# Making the query to find only the companies that do design

proy = {"_id":0,'offices': 1, "name": 1,'category_code':1}

query = {'category_code':'design'}

selected_offices = list(companies.find(query,proy))

# I am choosing only USA because it is the country with the most starbucks per capita after Monaco (which is not in the database) [see information here](https://www.silverdoorapartments.com/blog/which-country-has-the-most-starbucks-per-1000000-inhabitants/)

queryusa = {'$and':[{'category_code':'design'},{'offices.country_code':'USA'}]}

onlyusa = list(companies.find(queryusa,proy))
onlyusa

There is a company (99 designs) that has some branches outside USA, so I am removing them from the list

onlyusa[0]['offices']=onlyusa[0]['offices'][0]

onlyusa
