==> notebooks/2_info_from_API.ipynb <==
import sys
sys.path.append('../')

import os
from dotenv import load_dotenv
import requests
import json
import time
from functools import reduce
import operator
import pandas as pd
import src.functions as f

The list of cities I got from the database filtering is the following:

cities = ['San Francisco, CA, USA', 'Ellensburg, WA, USA', 'Brooklyn, NY, USA']

Making the names of the cities shorter

cities_short = [i.split(",")[0] for i in cities]

Creating a folder called "newdata" where I will save all my processed dataframes

if not os.path.exists(f"../newdata"):
    os.mkdir(f"../newdata")

Preparing the .env file, where I saved the 4square API tokens.

load_dotenv()

client_id = os.getenv("tok1")
client_secret = os.getenv("tok2")

The coordinates below are the ones from the city center of the cities I got from filtering the database.

sanfrancisco = {'type': 'Point', 'coordinates':[37.778480215129015, -122.40559118863916]}
ellensburg =  {'type': 'Point', 'coordinates':[46.996517823587766, -120.54550769638885]}
brooklyn =  {'type': 'Point', 'coordinates':[40.692147496548735, -73.98495174207916]}

These are the queries I want to make in order to find how far the cities coordinates are from these locations.

queries = ["vegan", "starbucks", "school","train station subway", "basketball court"]

def getFromDict(diccionario,mapa):
    return reduce (operator.getitem,mapa,diccionario)

def get_data (latitude, longitude, query):
    d = {}
    time.sleep(2)
    parametres = {"client_id" : client_id,
              "client_secret" : client_secret,
              "v": "20180323",
              "ll": f"{latitude},{longitude}", 
              "query":query,
              "limit": 100}
    time.sleep(1)

    resp = requests.get(url= 'https://api.foursquare.com/v2/venues/search', params=parametres).json()

    return resp

def respven(dic):
    return dic['response']['venues']

def extraetodo(json):
    todo = {"name": ["name"], "lat": ["location", "lat"], "lon": ["location", "lng"], "distance":['location', 'distance']} 
    total = []
    for elemento in json:
        dic = {key: getFromDict(elemento, value) for key,value in todo.items()}
        dic["location"] = {'type': 'Point', 'coordinates': [dic["lat"], dic["lon"]]}
        total.append(dic)
    return total

def fin(city, n):
    return extraetodo(respven(get_data(city['coordinates'][0], city['coordinates'][1], queries[n])))

def todf(fincity):
    fordf = []
    for dic in fincity:
            for i in dic:
                todf = {}
                todf["name"] = getFromDict(i,["name"])
                todf["lat"] = getFromDict(i,["lat"])
                todf["lon"] = getFromDict(i,["lon"])
                todf["location"] = getFromDict(i,["location"])
                todf["category"] = getFromDict(i,["category"])
                todf["distance"] = getFromDict (i,["distance"])
                fordf.append(todf)
    df = pd.DataFrame(fordf)
    return df

def dropbydist(df, dis, cat):
    return df.drop(df[(df.distance > dis) & (df.category == cat)].index, inplace=True)

These are the name of the data categories, which are based in my queries.

category = ["vegan_venues", "starbucks", "schools","train_stations", "basketball_courts" ]

# Making the API call and creating the cleaned dataframe from the API data

## San Fracisco

vegsanfrancisco = fin(sanfrancisco, 0)
for  i in  vegsanfrancisco:
    i['category']= 'vegan_venues'
    
stasanfrancisco = fin(sanfrancisco, 1)
for  i in  stasanfrancisco:
    i['category']= 'starbucks'
    
schsanfrancisco = fin(sanfrancisco, 2)
for  i in  schsanfrancisco:
    i['category']= 'schools'

trasanfrancisco = fin(sanfrancisco, 3)
for  i in  trasanfrancisco:
    i['category']= 'train_stations'

bassanfrancisco = fin(sanfrancisco, 4)
for  i in  bassanfrancisco:
    i['category']= 'basketball_courts'

finsanfrancisco = [vegsanfrancisco,stasanfrancisco, schsanfrancisco, trasanfrancisco, bassanfrancisco]

dfsanfrancisco = f.todf(finsanfrancisco)

dropbydist(dfsanfrancisco, 1500, category[0])
dropbydist(dfsanfrancisco, 1300, category[1])
dropbydist(dfsanfrancisco, 3000, category[2])
dropbydist(dfsanfrancisco, 2500, category[3])
dropbydist(dfsanfrancisco, 2000, category[4])

dfsanfrancisco.to_csv(f"../newdata/dfsanfran.csv", index = False)

## Ellensburg

vegell = fin(ellensburg, 0)
for  i in  vegell:
    i['category'] = 'vegan_venues'
    
staell = fin(ellensburg, 1)
for  i in  staell:
    i['category'] = 'starbucks'

schell = fin(ellensburg, 2)
for  i in  schell:
    i['category'] = 'schools'

traell = fin(ellensburg, 3)
for  i in  traell:
    i['category'] = 'train_stations'

basell = fin(ellensburg, 4)
for  i in  basell:
    i['category'] = 'basketball_courts'

finellensburg = [vegell, staell, schell, traell, basell]

dfellensburg = todf(finellensburg)

dropbydist(dfellensburg, 1500, category[0])
dropbydist(dfellensburg, 1300, category[1])
dropbydist(dfellensburg, 3000, category[2])
dropbydist(dfellensburg, 2500, category[3])
dropbydist(dfellensburg, 2000, category[4])

dfellensburg.to_csv(f"../newdata/dfellensburg.csv", index = False)

## Brooklyn

vegbro = fin(brooklyn, 0)
for  i in  vegbro:
    i['category'] = 'vegan_venues'

stabro = fin(brooklyn, 1)
for  i in  stabro:
    i['category'] = 'starbucks'
    
schbro = fin(brooklyn, 2)
for  i in  schbro:
    i['category'] = 'schools'
    
trabro = fin(brooklyn, 3)
for  i in  trabro:
    i['category'] = 'train_stations'

basbro = fin(brooklyn, 4)
for  i in  basbro:
    i['category'] = 'basketball_courts'

finbrooklyn = [vegbro, stabro, schbro, trabro, basbro]

dfbrooklyn = todf(finbrooklyn)

dropbydist(dfbrooklyn, 1500, category[0])
dropbydist(dfbrooklyn, 1300, category[1])
dropbydist(dfbrooklyn, 3000, category[2])
dropbydist(dfbrooklyn, 2500, category[3])
dropbydist(dfbrooklyn, 2000, category[4])

dfbrooklyn.to_csv(f"../newdata/dfbrooklyn.csv", index = False)
