==> notebooks/3_geoqueries.ipynb <==
from pymongo import MongoClient
from pymongo import GEOSPHERE
import geopandas as gpd
import pandas as pd
import shapely

client = MongoClient("localhost:27017")
db = client.get_database("ironhack")
db

# Getting dataframes per city

dfsanfran = pd.read_csv('../newdata/dfsanfran.csv')
dfell = pd.read_csv('../newdata/dfellensburg.csv')
dfbro = pd.read_csv('../newdata/dfbrooklyn.csv')

# Getting mongo collections per city

sf = db.get_collection("sanfrancisco")
ell = db.get_collection("ellensburg")
bro = db.get_collection("brooklyn")

# Making geodataframes

geosf = gpd.GeoDataFrame(dfsanfran, geometry=gpd.points_from_xy(dfsanfran.lon, dfsanfran.lat))
geosf.drop("location",axis=1, inplace=True)

geoell = gpd.GeoDataFrame(dfell, geometry=gpd.points_from_xy(dfell.lon, dfell.lat))
geoell.drop("location",axis=1, inplace=True)

geobro = gpd.GeoDataFrame(dfbro, geometry=gpd.points_from_xy(dfbro.lon, dfbro.lat))
geobro.drop("location",axis=1, inplace=True)

# Creating indexes

## San Francisco

geosf['geometry']=geosf['geometry'].apply(lambda x:shapely.geometry.mapping(x))
geosfdic = geosf.to_dict(orient="records")
sf.create_index([("geometry", GEOSPHERE)])
sf.insert_many(geosfdic)

## Ellensburg

geoell['geometry']=geoell['geometry'].apply(lambda x:shapely.geometry.mapping(x))
geoelldic = geoell.to_dict(orient="records")
ell.create_index([("geometry", GEOSPHERE)])
ell.insert_many(geoelldic)

## Brooklyn

geobro['geometry']=geobro['geometry'].apply(lambda x:shapely.geometry.mapping(x))
geobrodic = geobro.to_dict(orient="records")
bro.create_index([("geometry", GEOSPHERE)])
bro.insert_many(geobrodic)

# Making the Mongo gequeries to  get the distance from the cities' coordinates to the targetted locations

sanfrancisco = {'type': 'Point', 'coordinates':[-122.40559118863916, 37.778480215129015]}
ellensburg =  {'type': 'Point', 'coordinates':[-120.54550769638885, 46.996517823587766]}
brooklyn =  {'type': 'Point', 'coordinates':[40.692147496548735, -73.98495174207916]}

metres = 3000

## San Francisco geoquery

querysf = [{"$geoNear": {
            "near": sanfrancisco['coordinates'],
            "distanceField": "geodist",
            "maxDistance": 3000,
            "distanceMultiplier" :6371,
            "spherical": True}}]
geoloc_sf = list(sf.aggregate(querysf))

geoloc_sf = pd.DataFrame(geoloc_sf)
geoloc_sf.drop(['_id','distance','geometry'], axis=1, inplace=True)

geoloc_sf.to_csv(f"../newdata/dfgeosf.csv", index = False)

## Ellensburg geoquery

queryell = [{"$geoNear": {
            "near": ellensburg['coordinates'],
            "distanceField": "geodist",
            "maxDistance": 3000,
            "distanceMultiplier" :6371,
            "spherical": True}}]
geloc_ell = list(ell.aggregate(queryell))

geloc_ell = pd.DataFrame(geloc_ell)
geloc_ell.drop(['_id','distance','geometry'], axis=1, inplace=True)

geloc_ell.to_csv(f"../newdata/dfgeoell.csv", index = False)

## Brooklyn geoquery

querybro = [{"$geoNear": {
            "near": brooklyn['coordinates'],
            "distanceField": "geodist",
            "maxDistance": 3000,
            "distanceMultiplier" :6371,
            "spherical": True}}]
geoloc_bro = list(bro.aggregate(querybro))

geoloc_bro = pd.DataFrame(geoloc_bro)
geoloc_bro.drop(['_id','distance','geometry'], axis=1, inplace=True)

geoloc_bro['geodist'] = geoloc_bro['geodist']/100000
geoloc_bro

geoloc_bro.to_csv(f"../newdata/dfgeobro.csv", index = False)
