==> notebooks/4_geoweights.ipynb <==
import pandas as pd

# Specifying the weights per location category

marks = {"vegan_venues": 0.20, "starbucks":0.20, "schools":0.20, "train_stations":0.25, "basketball_courts":0.15} 

# Loading the dataframes from the Mongo geoqueries

dfsanfran = pd.read_csv("../newdata/dfgeosf.csv")
dfell = pd.read_csv('../newdata/dfgeoell.csv')
dfbro = pd.read_csv('../newdata/dfgeobro.csv')

# Making a new column with the punctuation per location (weights*distance)

def punct(df):  
    o = []
    for i,row in df.iterrows():
        if row['category'] == "vegan_venues":
            o.append(row['geodist']*marks['vegan_venues'])

        elif row['category'] == "starbucks":
            o.append(row['geodist']*marks['starbucks'])

        elif row['category'] == "schools":
            o.append(row['geodist']*marks['schools'])

        elif row['category'] == "train_stations":
            o.append(row['geodist']*marks['train_stations'])

        elif row['category'] == "basketball_courts":
            o.append(row['geodist']*marks['basketball_courts'])
    df['punctuation'] = o
    return df['punctuation']

punct(dfsanfran)
punct(dfell)
punct(dfbro)

dfsanfran

# Normalizing the data

## San Francisco

def normalization(df, col):
    df['normal'] = (df[col] - df[col].min()) / (df[col].max() - df[col].min())     
    df = df.reset_index(drop = False)
    return df

pundfsanfran = normalization(dfsanfran,'geodist')
pundfsanfran = dfsanfran.groupby('category').agg({'normal':'mean'}).reset_index()

pundfsanfran

meansanfran = sum([i for i in pundfsanfran['normal']])/len(pundfsanfran['normal'])
meansanfran

## Ellensburg

pundfell = normalization(dfell,'geodist')
pundfell = dfell.groupby('category').agg({'normal':'mean'}).reset_index()

pundfell

meanell = sum([i for i in pundfell['normal']])/len(pundfell['normal'])
meanell

## Brooklyn

pundfbro = normalization(dfbro,'geodist')
pundfbro = dfbro.groupby('category').agg({'normal':'mean'}).reset_index()

pundfbro

meanbrooklyn = sum([i for i in pundfbro['normal']])/len(pundfbro['normal'])
meanbrooklyn

# Calculating the mean values of the different cities

means = [meansanfran, meanell, meanbrooklyn]
col = ['San Francisco', 'Ellensburg', 'Brooklyn']
finalmeans = pd.DataFrame(means, index = col).T
finalmeans

 <font size="5"> Taking into account the studies, the best city to start our new company is San Francisco, followed by Brooklyn (as Ellensburg doesn't have any vegan restaurant and that is one of the requirements).</font> 

finalmeans.to_csv("../newdata/finalmeans.csv")
